moved intial movies call and state into Home component, renamed some variables for semantic cleanliness
