+++ web/film-club-client/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { Route, Switch } from 'react-router-dom';
 
 // Import Custom Components
@@ -7,25 +7,15 @@ import FilmDetailPage from './components/FilmDetailPage'
 
 
 function App() {
-  // const [showSidebar, setShowSidebar] = useState(false)
-  const [movies, setMovies] = useState([])
-  const [film, setFilm] = useState({})
-
-  // Get all film data and concat to 'movies' array for Infinite Scroll use
-  const getUpcomingMovies = async () => {
-      const page = (movies.length / 20) + 1
-      const response = await fetch(`/films/upcoming?page=${page}`)
-      const data = await response.json()
-      setMovies(movies.concat(data.films))
-  }
+  const [movieId, setMovieId] = useState({})
 
   return (
       <Switch>
         <Route exact path='/'>
-          <Home setFilm={setFilm} getUpcomingMovies={getUpcomingMovies} movies={movies} />
+          <Home setMovieId={setMovieId} />
         </Route>
         <Route path={'/detail/:id'}>
-          <FilmDetailPage film={film} />
+          <FilmDetailPage movieId={movieId} />
         </Route>
       </Switch>
   );
